Update col-span since it doesn't work
Locally using col-span didn't work using tailwind 3, so I propose we change this so it does work. Also there was a space in the class while using underscores on the other part of the string. 

Although tailwind 3 says you can use spaces, it's clearer to make it either all spaces or all underscores.